Move Telegram client code into separate module
