+++ lib/main.js
@@ -2,6 +2,7 @@ var restify = require("restify");
 var clients = require("restify-clients")
 var querystring = require("querystring");
 var {bold, link} = require("./format");
+var {sendMessage} = require("./telegram");
 
 var BOT_API_KEY = process.env.BOT_API_KEY;
 var TELEGRAM_API_KEY = process.env.TELEGRAM_API_KEY;
@@ -160,26 +161,3 @@ function handlePullRequestMerged(req, res) {
   sendMessage(message);
 }
 
-function sendMessage(text) {
-  console.log("Message to be set: " + text);
-
-  var client = clients.createClient({url: TELEGRAM_API_URL, headers: {"Content-Type": "application/x-www-form-urlencoded"}});
-  client.post("/bot" + TELEGRAM_API_KEY + "/sendMessage", function(err, req) {
-    req.on('result', function(err, res) {
-      res.body = '';
-      //res.setEncoding('utf8');
-      res.on('data', function(chunk) {
-        res.body += chunk;
-      });
-
-      res.on('end', function() {
-        console.log(res.body);
-      });
-    });
-
-    response = querystring.stringify({chat_id: TELEGRAM_CHAT_ID, text: text, parse_mode: "Markdown"});
-    console.log(response);
-    req.write(response);
-    req.end();
-  });
-}
